Bug: Dapper mapping failure - found with asserts
I had initially placed the asserations
Debug.Assert(conversation != null && conversation.ChatUserId == user.Id, "Conversation does not belong to user");
and lucky I did, because it allowed me to find a problem in the mapping
dapper to the fields. For some reason the select * was not mapping all the
fields of the structure. Solved by making the select explicit. Will do that¨
for all select * to avoid an issue.